+++ src/mapfile_parser/frontends/objdiff_report.py
@@ -9,17 +9,34 @@ import argparse
 import dataclasses
 import decomp_settings
 from pathlib import Path
+import os
 
 from .. import mapfile
+from .. import utils
+from ..internals import objdiff_report as report_internal
 
 
-def doObjdiffReport(mapPath: Path, outputPath: Path, prefixesToTrim: list[str], reportCategories: mapfile.ReportCategories, *, pathIndex: int=2, asmPath: Path|None=None, nonmatchingsPath: Path|None=None) -> int:
+def doObjdiffReport(
+        mapPath: Path,
+        outputPath: Path,
+        prefixesToTrim: list[str],
+        reportCategories: mapfile.ReportCategories,
+        *,
+        pathIndex: int=2,
+        asmPath: Path|None=None,
+        nonmatchingsPath: Path|None=None,
+        emitCategories: bool=False,
+        quiet: bool=False,
+    ) -> int:
     if not mapPath.exists():
         print(f"Could not find mapfile at '{mapPath}'")
         return 1
 
     mapFile = mapfile.MapFile()
     mapFile.readMapFile(mapPath)
+
+    if emitCategories:
+        printDefaultCategories(mapFile, prefixesToTrim)
 
     mapFile.writeObjdiffReportToFile(
         outputPath,
@@ -30,7 +47,221 @@ def doObjdiffReport(mapPath: Path, outputPath: Path, prefixesToTrim: list[str], 
         nonmatchingsPath=nonmatchingsPath,
     )
 
+    if not quiet:
+        report = report_internal.Report.readFile(outputPath)
+        if report is None:
+            utils.eprint(f"Unable to read back the generated report at {outputPath}")
+            return 1
+        table = report.asTableStr(sort=True)
+        print(table, end="")
+
+        # Output to GitHub Actions job summary, if available
+        summary_path = os.getenv("GITHUB_STEP_SUMMARY")
+        if summary_path is not None:
+            with open(summary_path, "a", encoding="UTF-8") as summary_file:
+                summary_file.write("```\n")
+                summary_file.write(table)
+                summary_file.write("```\n")
+
     return 0
+
+def printDefaultCategories(
+        mapFile: mapfile.MapFile,
+        prefixesToTrim: list[str],
+    ):
+    if len(prefixesToTrim) == 0:
+        # Manage a list of defaults
+        prefixesToTrim = [
+            "build/lib/",
+            "build/src/",
+            "build/asm/data/",
+            "build/asm/",
+            "build/",
+        ]
+
+    def removeSuffixes(path: Path) -> Path:
+        while path.suffix != "":
+            path = path.with_suffix("")
+        return path
+
+    def removePrefix(path: Path) -> Path:
+        current = str(path)
+        # Trim the first prefix found in the list
+        for x in prefixesToTrim:
+            if current.startswith(x):
+                current = current.removeprefix(x)
+                break
+        current = current.removeprefix("/")
+        return Path(current)
+
+    categoriesByPath: list[Category] = []
+    categoriesBySegment: list[Category] = []
+
+    def addCategoryPath(categoriesByPath: list[Category], cat_path: str):
+        ide = cat_path.strip("/")
+        for x in categoriesByPath:
+            if x.ide == ide:
+                return
+        cat = Category(ide, ide, [cat_path])
+        categoriesByPath.append(cat)
+
+    def addCategorySegment(categoriesByPath: list[Category], segmentName: str, cat_path: str):
+        ide = f"segment_{segmentName}"
+        for x in categoriesByPath:
+            if x.ide == ide:
+                if cat_path not in x.paths:
+                    x.paths.append(cat_path)
+                return
+        cat = Category(ide, f"Segment {segmentName}", [cat_path])
+        categoriesByPath.append(cat)
+
+    def reduceCategorySegmentPaths(categoriesByPath: list[Category]):
+        """
+        We want to reduce the list of paths as much as possible by reducing
+        them to their prefixes.
+        We want to convert a list of paths like this:
+        ```
+        - main/audio/sound
+        - sys/gtl
+        - sys/ml
+        - main/sfxlimit
+        - main/file
+        - sys/vi
+        - sys/rdp_reset
+        - main/title
+        - main/menu
+        - sys/om
+        - ultralib/src/audio/cspsetbank
+        - ultralib/src/audio/cspsetpriority
+        - ultralib/src/audio/sndpstop
+        - ultralib/src/audio/cseq
+        - game
+        ```
+        To a reduced list like this:
+        ```
+        - main
+        - sys
+        - ultralib
+        - game
+        ```
+
+        But the reduced list must not overlap paths from other segments.
+        """
+
+        # Ensure the given shortest path is not present in any other segment
+        def isUnique(categoriesByPath: list[Category], shortest: Path, currentIde: str) -> bool:
+            for cat2 in categoriesByPath:
+                if currentIde == cat2.ide:
+                    continue
+                for path2 in cat2.paths:
+                    if shortest in Path(path2).parents:
+                        return False
+            return True
+
+        # Build the list of reduced prefixes per segment
+        uniques: dict[str, list[Path]] = {}
+        for cat in categoriesByPath:
+            if len(cat.paths) <= 1:
+                continue
+            currentList: list[Path] = []
+            # Check each path for its reduced value
+            for p in cat.paths:
+                parents = list(Path(p).parents)
+                # Omit the `.` element in parents and iterate from the smaller
+                # prefix until the largest.
+                for shortest in parents[:-1][::-1]:
+                    if shortest in currentList:
+                        break
+                    if isUnique(categoriesByPath, shortest, cat.ide):
+                        currentList.append(shortest)
+                        break
+            if len(currentList) != 0:
+                uniques[cat.ide] = currentList
+
+        # Decide if we want to keep the current path or if we want to replace
+        # it with a prefix
+        def decide(old: str, prefixes: list[Path]) -> str:
+            parents = Path(old).parents
+            for prefix in prefixes:
+                if prefix in parents:
+                    return str(prefix)
+            return old
+
+        # Update paths with their prefixes for each segment
+        for ide, prefixes in uniques.items():
+            for cat in categoriesByPath:
+                if cat.ide == ide:
+                    newPaths = []
+                    for path in cat.paths:
+                        p = decide(path, prefixes)
+                        # Avoid duplication
+                        if p not in newPaths:
+                            newPaths.append(p)
+                    cat.paths = newPaths
+                    break
+
+    realSegmentsSeen = False
+    for segment in mapFile:
+        if segment.vrom is None and realSegmentsSeen:
+            # Usually debug sections
+            continue
+        if segment.vrom is not None:
+            realSegmentsSeen = True
+        for section in segment:
+            if section.isNoloadSection:
+                continue
+            suffixless = removeSuffixes(section.filepath)
+            prefixless = removePrefix(suffixless)
+
+            parts = prefixless.parts
+            if len(parts) > 1:
+                # Folder
+                cat_path = str(parts[0]) + "/"
+            elif len(parts) > 0:
+                # Top-level file
+                cat_path = str(parts[0])
+            else:
+                # Huh?
+                cat_path = "root"
+
+            addCategoryPath(categoriesByPath, cat_path)
+            addCategorySegment(categoriesBySegment, segment.name, str(prefixless))
+    reduceCategorySegmentPaths(categoriesBySegment)
+
+    print("""\
+tools:
+  mapfile_parser:
+    progress_report:
+      # output: report.json # Optional
+      check_asm_paths: True
+      # Change if the asm path in the build folder is deeper than two subfolders.
+      # i.e.: "build/us/asm/header.o" -> `path_index: 3`.
+      # i.e.: "build/us/asm/us/header.o" -> `path_index: 4`.
+      # path_index: 2
+      prefixes_to_trim:
+""", end="")
+    for trim in prefixesToTrim:
+        print(f"        - \"{trim}\"")
+
+    def printCategories(categories: list[Category]):
+        for cat in categories:
+            print(f"""\
+        - id: "{cat.ide}"
+          name: "{cat.name}"
+          paths:
+""", end="")
+            for p in cat.paths:
+                print(f"""\
+            - "{p}"
+""", end="")
+
+    print("      categories:")
+    print("        # Categories by path")
+    printCategories(categoriesByPath)
+    print()
+    print("        # Categories by segment")
+    printCategories(categoriesBySegment)
+
 
 def processArguments(args: argparse.Namespace, decompConfig: decomp_settings.Config|None=None):
     reportCategories = mapfile.ReportCategories()
@@ -91,15 +322,40 @@ def processArguments(args: argparse.Namespace, decompConfig: decomp_settings.Con
         asmPath = args.asmpath
         nonmatchingsPath = args.nonmatchingspath
 
-    exit(doObjdiffReport(mapPath, outputPath, prefixesToTrim, reportCategories, asmPath=asmPath, pathIndex=pathIndex, nonmatchingsPath=nonmatchingsPath))
+    emitCategories: bool = args.emit_categories
+    quiet: bool= args.quiet
+
+    exit(doObjdiffReport(
+        mapPath,
+        outputPath,
+        prefixesToTrim,
+        reportCategories,
+        asmPath=asmPath,
+        pathIndex=pathIndex,
+        nonmatchingsPath=nonmatchingsPath,
+        emitCategories=emitCategories,
+        quiet=quiet
+    ))
 
 def addSubparser(subparser: argparse._SubParsersAction[argparse.ArgumentParser], decompConfig: decomp_settings.Config|None=None):
     epilog = """\
 Visit https://decomp.dev/ and https://wiki.decomp.dev/tools/decomp-dev for more
 information about uploading the generated progress report.
 
+A summary of the generated progress report will be printed to stdout by
+default. Also this script will try to detect if it is running on a Github
+Action and print the same summary as a step summary. Use the `--quiet` flag to
+avoid this behaviour.
+
 This utility has support for a special section on the `decomp.yaml` file, which
 allows to avoid passing many arguments to utility.
+
+Use the `--emit-categories` flag to print categories generated automatically
+from your mapfile, using the `decomp.yaml` format.
+You are expected to tweak the generated categories to accomodate to your liking
+instead of using them as-is.
+You may also want to set `prefixes_to_trim` before using this flag, to improve
+the generation of these categories.
 
 Here's an example for this entry:
 
@@ -187,6 +443,9 @@ tools:
     if emitPathIndex:
         parser.add_argument("-i", "--path-index", help="Specify the index to start reading the file paths. Defaults to 2", type=int)
 
+    parser.add_argument("--emit-categories", help="Print automatically-generated categories from your mapfile, using the decomp.yaml format. These categories are expected to be tweaked and not used as-is.", action="store_true")
+    parser.add_argument("--quiet", help="Avoid printing the progress report to the stdout and to the Github action summary.", action="store_true")
+
     parser.set_defaults(func=processArguments)
 
 
@@ -199,7 +458,7 @@ class SpecificSettings:
     checkAsmPaths: bool
 
     @staticmethod
-    def fromDecompConfig(decompConfig) -> SpecificSettings|None:
+    def fromDecompConfig(decompConfig: decomp_settings.Config|None=None) -> SpecificSettings|None:
         if decompConfig is None:
             return None
 
@@ -254,19 +513,19 @@ class Category:
         ide = data.get("id")
         if ide is None:
             return None
-        assert isinstance(ide, str)
+        assert isinstance(ide, str), f"{type(ide)} {ide}"
 
         name = data.get("name")
         if name is None:
             return None
-        assert isinstance(name, str)
+        assert isinstance(name, str), f"{type(name)} {name}"
 
         paths = data.get("paths")
         if paths is None:
             return None
-        assert isinstance(paths, list)
+        assert isinstance(paths, list), f"{type(paths)} {paths}"
         for x in paths:
-            assert isinstance(x, str)
+            assert isinstance(x, str), f"{type(x)} {x}"
 
         return Category(
             ide,